requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
django-filter>=23.0
drf-yasg>=1.21.0
psycopg2-binary>=2.9.0
whitenoise>=6.6.0
gunicorn>=21.2.0
dj-database-url>=2.1.0

